refactor: centralize send to queue function
