--- apps/auto-email-sender/email-pre-renderer/src/sendToQueue.ts
@@ -1,16 +1,0 @@
-import { Channel } from 'amqplib'
-import { EmailQueues } from 'shared'
-
-type Props = Record<
-  string,
-  {
-    footerHTML: string
-    headerHTML: string
-    roles: string[]
-  }
->
-export async function sendToQueue(queueChannel: Channel, props: Props) {
-  const queue = EmailQueues.EmailPreRenderer
-  await queueChannel.assertQueue(queue)
-  queueChannel.sendToQueue(queue, Buffer.from(JSON.stringify(props)))
-}
