+++ apps/auto-email-sender/subs-to-queue/src/writeToQueues.ts
@@ -1,7 +1,7 @@
 import { Channel } from 'amqplib'
 import { SupabaseTable } from 'db/src/supabase/utilityTypes'
 import { EmailQueues } from 'shared'
-import { sendToQueue } from './sendToQueue'
+import { sendToQueue } from 'shared/src/queue/sendToQueue'
 
 export type Subscriber = Pick<
   SupabaseTable<'Subscribers'>,
